requirements.txt
fastapi
uvicorn[standard]
pypdf
sentence-transformers
faiss-cpu
numpy
python-multipart
aiofiles
requests
# LangChain + vector store dependencies
langchain
langchain-community
langchain-openai
chromadb
openai
# deployment
python-dotenv
